+++ iis-workflows/iis-workflows-top/src/main/resources/eu/dnetlib/iis/workflows/top/common/export/job.properties
@@ -15,8 +15,8 @@ input_aux_dataset_existing_id=/share/import/datasetid/empty
 #export
 action_hbase_table_initialize=true
 action_hbase_table_name=db_actions_xxx
-action_set_id=xxx
-action_set_id_entity_dataset=xxx
+action_set_id=generic_actionset_id
+action_set_id_entity_dataset=entity_dataset_actionset_id
 
 #required when exporting to external cluster
 #action_hbase_remote_zookeeper_quorum=jobtracker.t.hadoop.research-infrastructures.eu:2182,quorum1.t.hadoop.research-infrastructures.eu:2182,quorum2.t.hadoop.research-infrastructures.eu:2182,quorum3.t.hadoop.research-infrastructures.eu:2182,quorum4.t.hadoop.research-infrastructures.eu:2182
@@ -27,7 +27,7 @@ input_document_to_mdstore=/user/marek.horst/importer/dataset/working_dir/out_dat
 input_document_to_dataset=/share/referenceextraction/dataset/puma/2014-06-27
 
 #hbase export mode when output set to $UNDEFINED$ sequence file export mode when output dir provided
-#output_remote_location=hdfs://iis-cluster-nn/tmp/
+output_remote_location=hdfs://iis-cluster-nn/tmp/iis_export/primary
 
 #need to disable fork-join test due to 'decision' related problem causing: Error: E0743 : E0743: Multiple "ok to" transitions to the same node, [transformers_common_union_document_text], are not allowed
 oozie.wf.validate.ForkJoin=false
